requirements.txt
gradio==4.21.0
python-dotenv==1.0.1
requests==2.31.0